refactor(cli): split main into srp modules
