+++ crates/uselesskey-cli/src/srp/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod args;
+pub(crate) mod artifact_io;
+pub(crate) mod profile_docs;
